refactor: consolidate fallback helpers
